changed titles color
